requirements.txt
fastapi
uvicorn
python-multipart
c2pa-python==0.27.1
aiohttp
pydantic
pydantic-settings
runpod
python-dotenv
opencv-python-headless
Pillow
numpy
pandas
plotly
google-genai
firebase-admin
upstash-redis
streamlit>=1.40.0
protobuf>=4.25.0
simple-lama-inpainting
pillow-heif
psutil
sentry-sdk
python-json-logger
axiom-py